Audio: Fixed Audio Issues and Detune not applied correctly!

+++ Plugin/Source/VeNo/Sound/Generation/Oscillator.h
@@ -9,16 +9,16 @@ class Oscillator {
 public:
   static bool setup(OscillatorData &, size_t instanceId);
   static bool prepare(OscillatorData &);
-  static bool process(OscillatorData &, SingleVoiceData &, int currentNote,
+  static bool process(OscillatorData &, SingleVoiceData &, float currentNote,
                       double sR);
   static bool finish(OscillatorData &, SingleVoiceData &);
   static void prepareVoice(OscillatorData &osc, SingleVoiceData voice);
-  static float renderVoice(SingleVoiceData &voice, DetuneState &detune,
+  static double renderVoice(SingleVoiceData &voice, DetuneState &detune,
                            double inc, const Wave &table, int idx);
 
 protected:
   static void updateFrequency(OscillatorData &, SingleVoiceData &,
-                              int currentNote);
+                              float currentNote);
   static void render(OscillatorData &, SingleVoiceData &, double sR);
   Oscillator() = default;
 };
