+++ Plugin/Source/VeNo/Sound/Tables/TableCreator.h
@@ -12,8 +12,8 @@ protected:
   static void CreateTable(int waveTableIndex, const std::function<void(RawTable*)>& generateWave);
   static void FillTables(RawTable&, WaveTable*);
   static void CreateTableRaw(RawTable&, WaveTable*);
-  static void MakeWaveTable(Wave &, double topFreq, float *tempData);
-  static void FFT(size_t i, float *ar, float *ai);
+  static void MakeWaveTable(Wave &, double topFreq, double *tempData);
+  static void FFT(int i, double *ar, double *ai);
   static double FindScale(Wave& wave);
   static void ApplyCutoff(Wave& wave, double topFreq);
